pyproject.toml
[project]
name = "telegram-ai-agent"
version = "0.1.0"
description = "Telegram bot for running Claude Code or Codex agents from chat"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "aiogram>=3.15",
    "pydantic-settings>=2.7",
    "deepgram-sdk>=5.0",
    "aiohttp>=3.9",
    "aiofiles>=24.1",
    "pyyaml>=6.0.3",
    "mcp>=1.0",
    "httpx>=0.25",
]

[project.scripts]
telegram-bot = "telegram_bot.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
    "mypy>=1.14",
    "types-aiofiles>=24.1",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
# messages.py holds translations — Cyrillic chars are intentional
"src/telegram_bot/core/messages.py" = ["RUF001"]
# These files intentionally match Claude/Codex TUI glyphs rendered by the CLI.
"src/telegram_bot/core/tui/capture.py" = ["RUF001", "RUF003"]
"src/telegram_bot/core/tui/modal_detect.py" = ["RUF001", "RUF002", "RUF003"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram_bot"]
